+++ demo/app/demo-main/index.js
@@ -157,7 +157,6 @@ export default class App extends Component {
         groups={groups}
         items={items}
         keys={keys}
-        fixedHeader="fixed"
         sidebarWidth={150}
         sidebarContent={<div>Above The Left</div>}
         // rightSidebarWidth={150}