+++ demo/app/demo-sticky-header/index.js
@@ -56,7 +56,6 @@ export default class App extends Component {
           groups={groups}
           items={items}
           keys={keys}
-          fixedHeader="sticky"
           sidebarWidth={150}
           sidebarContent={<div>Above The Left</div>}
           rightSidebarWidth={150}
@@ -105,7 +104,6 @@ export default class App extends Component {
           groups={groups}
           items={items}
           keys={keys}
-          fixedHeader="sticky"
           stickyOffset={100}
           sidebarWidth={150}
           sidebarContent={<div>Above The Left</div>}