+++ src/lib/lines/VerticalLines.js
@@ -12,13 +12,7 @@ export default class VerticalLines extends Component {
     lineCount: PropTypes.number.isRequired,
     minUnit: PropTypes.string.isRequired,
     timeSteps: PropTypes.object.isRequired,
-    fixedHeader: PropTypes.string.isRequired,
     height: PropTypes.number.isRequired
-  }
-
-  static defaultProps = {
-    fixedHeader: 'sticky',
-    dayBackground: null
   }
 
   shouldComponentUpdate(nextProps) {
@@ -30,7 +24,6 @@ export default class VerticalLines extends Component {
       nextProps.lineCount === this.props.lineCount &&
       nextProps.minUnit === this.props.minUnit &&
       nextProps.timeSteps === this.props.timeSteps &&
-      nextProps.fixedHeader === this.props.fixedHeader &&
       nextProps.height === this.props.height
     )
   }
@@ -60,8 +53,7 @@ export default class VerticalLines extends Component {
         const lineWidth = firstOfType ? 2 : 1
         const labelWidth =
           Math.ceil((nextTime.valueOf() - time.valueOf()) * ratio) - lineWidth
-        const leftPush =
-          this.props.fixedHeader !== 'none' && firstOfType ? -1 : 0
+        const leftPush = firstOfType ? -1 : 0
 
         const classNames =
           'rct-vl' +